Add CreatePlaylist && etc
Move main back to`mp3Player.java`
