+++ src/CreatePlaylist.java
@@ -1,0 +1,8 @@
+import javax.swing.*;
+
+public class CreatePlaylist {
+    private JPanel mainPanel;
+    private JPanel namePanel;
+    private JPanel descriptionPanel;
+    private JPanel buttonPanel;
+}
